support DifferentialDownloader in macOS
